+++ src/test/java/com/epam/reportportal/junit/features/suites/SuiteOfSuitesClass.java
@@ -16,8 +16,6 @@
 
 package com.epam.reportportal.junit.features.suites;
 
-import com.epam.reportportal.junit.features.suites.tests.FirstSuiteTest;
-import com.epam.reportportal.junit.features.suites.tests.SecondSuiteTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;